Working test server implemented
Installed Node.js on my computer and created a test server using Express.js framework and http requesting. All of this is under the Joe_test folder. All Node.js packages are under the config folder, not sure if that is an accurate name, but refer to that folder when accessing them.